// src/enums/migration-mode.ts
// -----------------------------------------------------------------------------------------
// #region Enums
// -----------------------------------------------------------------------------------------

/**
 * Enum representing each Entity Framework migration command operation
 */
enum MigrationMode {
    ADD = "ADD",
    DELETE = "DELETE",
    RUN = "RUN",
}

// #endregion Enums

// -----------------------------------------------------------------------------------------
// #region Exports
// -----------------------------------------------------------------------------------------

export { MigrationMode };

// #endregion Exports